pyproject.toml
[build-system]
requires = ["setuptools>=54", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "control-lab-ly"
version = "2.1.0"
description = "Lab Equipment Automation Package"
readme = "docs/README.md"
authors = [
    {name="Chang Jie Leong", email="changjie.leong@outlook.com"}
]
license = "MIT"
keywords = [
    "python",
    "lab automation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering"
]
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.9",
    "numpy>=2.0",
    "opencv_python>=4.10.0",
    "pandas>=2.0",
    "parse>=1.20",
    "pyserial>=3.5",
    "PyYAML>=6.0.1",
    "scipy>=1.14",
    "websockets>=15.0.1",
]

[project.urls]
Homepage = "https://kylejeanlewis.github.io/control-lab-ly/"
GitHub = "https://github.com/kylejeanlewis/control-lab-ly"
Documentation = "https://control-lab-ly.readthedocs.io/en/latest/"
Changelog = "https://github.com/kylejeanlewis/control-lab-ly/blob/main/docs/CHANGELOG.md"
Tracker = "https://github.com/kylejeanlewis/control-lab-ly/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=2.11",
    "pytest-rerunfailures>=15.0",
    "pytest-xdist>=3.6",
    "tox>=4.27"
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
    "mkdocstrings-python>=1.16.12",
    "ruff>=0.12.0"
]
all = [
    "pyModbusTCP>=0.2",
    "easy-biologic>=0.4.0",
    "nest-asyncio>=1.6.0",
    "setuptools>=71.0.3",
    "PyMeasure>=0.15"
]
ax8 = ["pyModbusTCP>=0.2"]
biologic = [
    "easy-biologic>=0.4.0",
    "nest-asyncio>=1.6.0",
    "setuptools>=71.0.3"
]
keithley = ["PyMeasure>=0.15"]

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
include = ["controllably*", "docs*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.png", "*.md", "*.rst"]


[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "socket: mark test socket connection"
]


[tool.tox]
legacy_tox_ini = """
[tox]
minversion = 3.10.0
envlist = py310, py311, py312
isolated_build = true

[gh-actions]
python =
    3.10: py310
    3.11: py311
    3.12: py312

[testenv]
setenv =
    PYTHONPATH = {toxinidir}
deps =
    -r {toxinidir}/dev/requirements_all.txt
    -r {toxinidir}/dev/requirements_dev.txt
commands =
    pytest --cov=controllably/core -n auto -m "not socket" --reruns 3 --reruns-delay 1 {posargs}
    pytest --cov=controllably/core -n 1 -m socket --reruns 3 --reruns-delay 1 {posargs}
"""

[tool.ruff.lint.per-file-ignores]
# Ignore F401 (unused imports) in all __init__.py files
"__init__.py" = ["F401"]
